[gameplay - Alpha 26]
Removed the ability of the Han to research 'Will to Fight' from a
captured Fortress. Reported by @faction_02 here:
https://wildfiregames.com/forum/topic/71578-feedbacks-from-a26-svn-tests/?do=findComment&comment=482636

Increased the cost and research time of 'Art of War' to match the other
two Han special Fortress techs.

On the backend moved the Han techs to the Fortress parent template.
Seemed cleaner.

This was SVN commit r26516.