[Han Chinese - Alpha 26] This retools the Imperial Court upgrade for the Civic Center.
Discussion:
https://wildfiregames.com/forum/topic/71582-fixing-chinese-han/

This removes the Champion training aspect and reduces the GarrisonHolder
from 2x to a 1.5x bonus, the same as the health and capture points
bonus. 2x garrisoning caused there to be a prohibitive number of arrows.

Replaces the champions with faster batch training speed. Unlocked heroes
still remain.

This was SVN commit r26634.