[maps] Adjustments to Fortress.js
Houses are placed on top of each other, metal availability is bad (and
the textures are old).

First two problems reported by Gurken Khan:
https://wildfiregames.com/forum/topic/82754-alpha-26-pre-releaserelease-candidate-build-testing/page/3/#comment-502958

Solution: use new balanced metal mine placement, update the textures and
replace the house switch/ the houses by a temple and a market.

An in-depth refactoring of the wall placement code should be made in the
future nevertheless.

accepted by: @lyv
Differential revision: https://code.wildfiregames.com/D4696
This was SVN commit r26940.