Fix victory conditions loading (fixes Hero Garrison in regicide crashloop).
VictoryConditions.js gamesetting sends multiple updates to 'active' and
'disabled', instead of just one single update at the end. This is
listened to by amongst others the regicide 'hero garrison' setting, thus
`HeroGarrison` is deactivated, and then tries to be reactivated from the
init attributes, and then disabled again later.

This implements a simple fix, VictoryConditions only sends a single
update in FromInitAttributes with the end-state.

Differential Revision: https://code.wildfiregames.com/D4793
This was SVN commit r27135.