Refactor Capturable Farmstead feature/templates
Attempts to address Invalid warning messages for capturable farmsteads
and creates a mixin as suggested by @Langbart

Renames the "farmsteads_" template files to "mill_" to eliminate
confusion with standard buildable farmsteads and changes the refs in the
skirmish maps to suit.

Genericizes the name to Rotary Mill instead of Farmstead, again to
eliminate confusion between the two (the former is a capturable map
object, while the latter is a player-buildable structure).

Fixes #6818
Differential revision: https://code.wildfiregames.com/D5113
This was SVN commit r27835.